Attach orbit controls after model is loaded
Orbit controls check the target to orbit around. In our code, we load
the target dynamically becacuse the url of the model is not static--it's
constructured in JS based on the model's id and auth id. As a result, we
need to attach the orbit controls after we load the target so it finds
it